requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
redis==5.0.1

# Data Processing
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3

# API Clients
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0

# LLM Integration
openai==1.10.0

# Task Queue
celery==5.3.6
celery[redis]==5.3.6

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2024.1
tabulate==0.9.0

# Monitoring & Logging
prometheus-client==0.19.0
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
